--- client/src/assets/dummyData.js
@@ -1,349 +1,0 @@
-const products = [
-  {
-    id: 0,
-    name: "Orvieto Classico",
-    country: "Italy",
-    type: "White/Dry",
-    price: "289",
-    rating: 5,
-    imageUrl: "/src/assets/photos/1.jpg",
-    year: 2021,
-    amount: 69,
-  },
-  {
-    id: 1,
-    name: "Monastrell Salina",
-    country: "Spain",
-    type: "Red/Dry",
-    price: "299",
-    rating: 3.5,
-    imageUrl: "/src/assets/photos/2.jpg",
-    year: 2022,
-    amount: 3,
-  },
-  {
-    id: 2,
-    name: "Château Margaux",
-    country: "France",
-    type: "Red/Dry",
-    price: "1299",
-    rating: 4.8,
-    imageUrl: "/src/assets/photos/3.jpg",
-    year: 2023,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 3,
-    name: "Riesling Kabinett",
-    country: "Germany",
-    type: "White/Sweet",
-    price: "189",
-    rating: 4.2,
-    imageUrl: "/src/assets/photos/4.jpg",
-    year: 2021,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 4,
-    name: "Malbec Reserva",
-    country: "Argentina",
-    type: "Red/Dry",
-    price: "239",
-    rating: 4.5,
-    imageUrl: "/src/assets/photos/5.jpg",
-    year: 2022,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 5,
-    name: "Sauvignon Blanc",
-    country: "New Zealand",
-    type: "White/Dry",
-    price: "179",
-    rating: 4.0,
-    imageUrl: "/src/assets/photos/6.jpg",
-    year: 2023,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 6,
-    name: "Zinfandel",
-    country: "United States",
-    type: "Red/Dry",
-    price: "199",
-    rating: 4.7,
-    imageUrl: "/src/assets/photos/7.jpg",
-    year: 2021,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 7,
-    name: "Rioja Reserva",
-    country: "Spain",
-    type: "Red/Dry",
-    price: "219",
-    rating: 4.4,
-    imageUrl: "/src/assets/photos/8.jpg",
-    year: 2022,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 8,
-    name: "Chardonnay",
-    country: "Australia",
-    type: "White/Oaked",
-    price: "259",
-    rating: 4.6,
-    imageUrl: "/src/assets/photos/9.jpg",
-    year: 2023,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 9,
-    name: "Pinot Noir",
-    country: "New Zealand",
-    type: "Red/Dry",
-    price: "299",
-    rating: 4.2,
-    imageUrl: "/src/assets/photos/10.jpg",
-    year: 2021,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 10,
-    name: "Sancerre",
-    country: "France",
-    type: "White/Dry",
-    price: "189",
-    rating: 4.8,
-    imageUrl: "/src/assets/photos/11.jpg",
-    year: 2022,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 11,
-    name: "Merlot",
-    country: "Chile",
-    type: "Red/Dry",
-    price: "179",
-    rating: 4.0,
-    imageUrl: "/src/assets/photos/12.jpg",
-    year: 2023,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 12,
-    name: "Cabernet Sauvignon",
-    country: "United States",
-    type: "Red/Dry",
-    price: "349",
-    rating: 4.9,
-    imageUrl: "/src/assets/photos/13.jpg",
-    year: 2021,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 13,
-    name: "Sauvignon Blanc Reserve",
-    country: "New Zealand",
-    type: "White/Dry",
-    price: "279",
-    rating: 4.7,
-    imageUrl: "/src/assets/photos/14.jpg",
-    year: 2022,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 14,
-    name: "Barolo Riserva",
-    country: "Italy",
-    type: "Red/Dry",
-    price: "589",
-    rating: 4.5,
-    imageUrl: "/src/assets/photos/1.jpg",
-    year: 2023,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 15,
-    name: "Chenin Blanc",
-    country: "South Africa",
-    type: "White/Dry",
-    price: "209",
-    rating: 4.2,
-    imageUrl: "/src/assets/photos/2.jpg",
-    year: 2021,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 16,
-    name: "Malbec Gran Reserva",
-    country: "Argentina",
-    type: "Red/Dry",
-    price: "449",
-    rating: 4.6,
-    imageUrl: "/src/assets/photos/3.jpg",
-    year: 2022,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 17,
-    name: "Syrah/Shiraz",
-    country: "Australia",
-    type: "Red/Dry",
-    price: "269",
-    rating: 4.3,
-    imageUrl: "/src/assets/photos/4.jpg",
-    year: 2023,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 18,
-    name: "Albariño",
-    country: "Spain",
-    type: "White/Dry",
-    price: "219",
-    rating: 4.1,
-    imageUrl: "/src/assets/photos/5.jpg",
-    year: 2021,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 19,
-    name: "Gewürztraminer",
-    country: "Germany",
-    type: "White/Sweet",
-    price: "189",
-    rating: 4.0,
-    imageUrl: "/src/assets/photos/6.jpg",
-    year: 2022,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 20,
-    name: "Cabernet Franc",
-    country: "France",
-    type: "Red/Dry",
-    price: "309",
-    rating: 4.5,
-    imageUrl: "/src/assets/photos/7.jpg",
-    year: 2023,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 21,
-    name: "Viognier",
-    country: "United States",
-    type: "White/Dry",
-    price: "239",
-    rating: 4.2,
-    imageUrl: "/src/assets/photos/8.jpg",
-    year: 2021,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 22,
-    name: "Pinotage",
-    country: "South Africa",
-    type: "Red/Dry",
-    price: "189",
-    rating: 4.0,
-    imageUrl: "/src/assets/photos/9.jpg",
-    year: 2022,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 23,
-    name: "Montepulciano d'Abruzzo",
-    country: "Italy",
-    type: "Red/Dry",
-    price: "199",
-    rating: 4.3,
-    imageUrl: "/src/assets/photos/10.jpg",
-    year: 2023,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 24,
-    name: "Chablis",
-    country: "France",
-    type: "White/Dry",
-    price: "259",
-    rating: 4.8,
-    imageUrl: "/src/assets/photos/11.jpg",
-    year: 2021,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 25,
-    name: "Merlot Reserva",
-    country: "Chile",
-    type: "Red/Dry",
-    price: "219",
-    rating: 4.6,
-    imageUrl: "/src/assets/photos/12.jpg",
-    year: 2022,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 26,
-    name: "Zweigelt",
-    country: "Austria",
-    type: "Red/Dry",
-    price: "229",
-    rating: 4.4,
-    imageUrl: "/src/assets/photos/13.jpg",
-    year: 2023,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 27,
-    name: "Sémillon",
-    country: "Australia",
-    type: "White/Dry",
-    price: "209",
-    rating: 4.2,
-    imageUrl: "/src/assets/photos/14.jpg",
-    year: 2021,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 28,
-    name: "Carmenère",
-    country: "Chile",
-    type: "Red/Dry",
-    price: "269",
-    rating: 4.7,
-    imageUrl: "/src/assets/photos/1.jpg",
-    year: 2022,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 29,
-    name: "Vermentino",
-    country: "Italy",
-    type: "White/Dry",
-    price: "189",
-    rating: 4.0,
-    imageUrl: "/src/assets/photos/2.jpg",
-    year: 2023,
-    amount: getRandomNumber(0, 100),
-  },
-  {
-    id: 30,
-    name: "Shiraz",
-    country: "Australia",
-    type: "Red/Dry",
-    price: "319",
-    rating: 4.9,
-    imageUrl: "/src/assets/photos/3.jpg",
-    year: 2021,
-    amount: getRandomNumber(0, 100),
-  },
-];
-
-function getRandomNumber(min, max) {
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-}
-
-export default products;
