--- client/src/assets/photos/README.md
@@ -1,1 +1,0 @@
-src: https://www.wine-express.pl/wino
